Simulate results from a single arm Bayesian Phase 2 Trial

Trial is a Bayesian phase 2, sequential trial as outlined in Thall, Peter F., and Richard Simon. "Practical Bayesian guidelines for phase IIB clinical trials." Biometrics (1994): 337-349.

ph2_single_bayes_seq_sim(
  omega_s_mu,
  omega_s_width,
  omega_s_w_conf_level = 0.9,
  omega_e_c,
  delta_0,
  n_min,
  n_max,
  sim_n = 1,
  pr_low = 0.05,
  pr_high = 0.95,
  mu_e = omega_s_mu + delta_0,
  verbose = FALSE,
  quiet = FALSE
)

Arguments

omega_s_mu
Mean of prior distribution on omega_s

omega_s_width
Width of the probability interval (set by `w_conf_level=`) running from the `(1 + w_conf_level) / 2` to the `(1 - w_conf_level) / 2` percentiles

omega_s_w_conf_level
Confidence level of width of probabilty specified

omega_e_c
Concentration parameter for prior distribution on omega_e

delta_0
Targeted improvement of treatment for treatment E over S

n_min
Minimum number of patients that may be enrolled

n_max
Maximum number of patients that may be enrolled

sim_n
Number of simulated trials to create

pr_low
Lower probability limit for concluding treatment not promising

pr_high
Upper probablity limit for concluding treatment promising

mu_e
True rate of success with experimental treatment in `omega_s_width`

verbose
When TRUE, additional information is returned as an attribute

quiet
Run with no notes, progress bars, etc.